Principal Component Analysis
In 8.10, a simple PCA is performed on NYSE data to determine the
proportion of variance explained by the principal components to
conclude if the data can be summarized in fewer than 5 dimensions. From
our PCA it looks like data can be summarized in 3 dimensions

In 8.11, PCA of Census Data is performed, correlation coefficients are
analyzed, and a comparison of PCA through covariance matrix and PCA
through correlation matrix is performed.

In 8.23, PCA without original data and only covariance matrix is
performed.